changes/requirements.txt
alabaster==0.7.10
amqp==2.2.2
anyjson==0.3.3
apiclient==1.0.3
argh==0.26.2
atelier==1.0.7
Babel==2.5.1
billiard==3.5.0.3
braintree==3.38.0
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.4.1
defusedxml==0.5.0
Django==1.11.5
django-admin-tools==0.8.1
django-allauth==0.33.0
django-appconf==1.0.2
django-async-messages==0.3.1
django-celery==3.2.1
django-common-helpers==0.9.1
django-countries==4.6.1
django-cron==0.5.0
django-daterange-filter==1.3.0
django-debug-toolbar==1.8
django-el-pagination==3.1.0
django-filter==1.0.4
django-fluent-dashboard==0.6.1
django-imagekit==4.0.1
django-jquery==3.1.0
django-jquery-lightbox==2.6
django-memcached==0.1.2
django-redis==4.8.0
django-rq==0.9.6
django-social-auth==0.7.28
django-social-share==1.1.2
django-tinymce==2.6.0
django-youtube==0.2
djangorestframework==3.6.4
docutils==0.14
easy-thumbnails==2.4.2
future==0.16.0
gdata==3.0.0
gitdb==0.6.4
gitdb2==2.0.2
GitPython==2.1.5
google-api-python-client==1.6.4
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
invoke==0.21.0
Jinja2==2.9.6
kombu==4.1.0
lxml==4.0.0
MarkupSafe==1.0
msgpack-python==0.4.8
oauth2==1.9.0.post1
oauth2client==4.1.2
oauthlib==2.0.4
olefile==0.44
pilkit==2.0
Pillow==4.2.1
psycopg2==2.7.3.1
py==1.4.34
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycrypto==2.6.1
pycryptodomex==3.4.7
Pygments==2.2.0
pyjwkest==1.3.6
PyJWT==1.5.3
pytest==3.2.2
pytest-cov==2.5.1
python-dateutil==2.6.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rq==0.8.2
rsa==3.4.2
simplejson==3.11.1
six==1.11.0
smmap==0.9.0
smmap2==2.0.3
snowballstemmer==1.2.1
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sockjs-tornado==1.0.3
-e git+https://github.com/jazzband/sorl-thumbnail.git@v12.4
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.3
SwampDragon==0.4.2.2
SwampDragon-auth==0.1.3
tlslite==0.4.9
tornado==4.5.2
tornado-redis==2.4.18
Unipath==1.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
